requirements.txt
python-telegram-bot==22.5
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
jdatetime>=5.0.0
pytz>=2024.1
Flask>=3.0.0
Flask-CORS>=4.0.0
qrcode>=7.4.2
Pillow>=10.0.0
mysql-connector-python>=8.2.0
gunicorn>=21.2.0
psutil>=5.9.0
